format: Allow virt-format test to be skipped by setting SKIP_TEST_VIRT_FORMAT_SH=1. (cherry picked from commit 101fd44fc5ce3055da91a5bf8aa0d89a5d2f3a79)

+++ format/test-virt-format.sh
@@ -20,6 +20,11 @@
 
 set -e
 
+if [ -n "$SKIP_TEST_VIRT_FORMAT_SH" ]; then
+    echo "$0: test skipped because environment variable is set."
+    exit 77
+fi
+
 rm -f test1.img
 
 ../fish/guestfish -N bootrootlv exit
